adapter: Migrate to Microsoft.VisualStudio.Shared.VSCodeDebugProtocol
This patch adopts the Microsoft-released DAP protocol library in favour
of the hand-rolled one in vscode-mono-debug. This allows us to
completely drop having to pull in code form vscode-mono-debug.

Note that support for process-discovery based attachment was disabled in
order to simplify the porting. The expectation is that this will be
dropped entirely in the future.